+++ Projects/Server/Items/Container.cs
@@ -792,30 +792,31 @@ namespace Server.Items
                     if (ParentsContain<BankBox>()) // Root Parent is the Mobile.  Parent could be another containter.
                     {
                         list.Add(
-                            1073841,
+                            1073841, // Contents: ~1_COUNT~/~2_MAXCOUNT~ items, ~3_WEIGHT~ stones
                             "{0}\t{1}\t{2}",
                             TotalItems,
                             MaxItems,
                             TotalWeight
-                        ); // Contents: ~1_COUNT~/~2_MAXCOUNT~ items, ~3_WEIGHT~ stones
+                        );
                     }
                     else
                     {
                         list.Add(
-                            1072241,
+                            1072241, // Contents: ~1_COUNT~/~2_MAXCOUNT~ items, ~3_WEIGHT~/~4_MAXWEIGHT~ stones
                             "{0}\t{1}\t{2}\t{3}",
                             TotalItems,
                             MaxItems,
                             TotalWeight,
                             MaxWeight
-                        ); // Contents: ~1_COUNT~/~2_MAXCOUNT~ items, ~3_WEIGHT~/~4_MAXWEIGHT~ stones
+                        );
                     }
 
                     // TODO: Where do the other clilocs come into play? 1073839 & 1073840?
                 }
                 else
                 {
-                    list.Add(1050044, "{0}\t{1}", TotalItems, TotalWeight); // ~1_COUNT~ items, ~2_WEIGHT~ stones
+                    // ~1_COUNT~ items, ~2_WEIGHT~ stones
+                    list.Add(1050044, "{0}\t{1}", TotalItems, TotalWeight);
                 }
             }
         }
