+++ Projects/UOContent/Engines/BulkOrders/LargeBODGump.cs
@@ -65,6 +65,7 @@ namespace Server.Engines.BulkOrders
 
             if (deed.Material != BulkMaterialType.None)
             {
+                // All items must be made with x material.
                 AddHtmlLocalized(
                     75,
                     y,
@@ -72,18 +73,19 @@ namespace Server.Engines.BulkOrders
                     20,
                     GetMaterialNumberFor(deed.Material),
                     0x7FFF
-                ); // All items must be made with x material.
+                );
             }
 
             AddButton(125, 168 + entries.Length * 24, 4005, 4007, 2);
+
             AddHtmlLocalized(
                 160,
                 168 + entries.Length * 24,
                 300,
                 20,
-                1045155,
+                1045155, // Combine this deed with another deed.
                 0x7FFF
-            ); // Combine this deed with another deed.
+            );
 
             AddButton(125, 192 + entries.Length * 24, 4005, 4007, 1);
             AddHtmlLocalized(160, 192 + entries.Length * 24, 120, 20, 1011441, 0x7FFF); // EXIT
