+++ Projects/UOContent/Engines/BulkOrders/SmallBOD.cs
@@ -128,9 +128,8 @@ namespace Server.Engines.BulkOrders
 
             if (m_AmountCur >= AmountMax)
             {
-                from.SendLocalizedMessage(
-                    1045166
-                ); // The maximum amount of requested items have already been combined to this deed.
+                // The maximum amount of requested items have already been combined to this deed.
+                from.SendLocalizedMessage(1045166);
             }
             else if (Type == null || objectType != Type && !objectType.IsSubclassOf(Type) ||
                      !(item is BaseWeapon) && !(item is BaseArmor) && !(item is BaseClothing))
