+++ Projects/UOContent/Engines/BulkOrders/SmallBODAcceptGump.cs
@@ -54,9 +54,9 @@ namespace Server.Engines.BulkOrders
                         deed.RequireExceptional ? 192 : 168,
                         350,
                         20,
-                        GetMaterialNumberFor(deed.Material),
+                        GetMaterialNumberFor(deed.Material), // All items must be made with x material.
                         0x7FFF
-                    ); // All items must be made with x material.
+                    );
                 }
             }
